4. About potions ||| Create Potion model, PotionRepository, PotionService, PotionController and RecipeService
a) Create Potion model. Potion has an id, name, a Student who brews it, a list of ingredients, a BrewingStatus, and a Recipe. Until a potion does not contain five ingredients, its BrewingStatus is brew. After that, if there is already a Recipe with the same ingredients (in any order), the status is replica. Otherwise, the status is discovery.
b) Create an endpoint at /potions. Here you must list all potions.
c) Potions are persisted in the database.