5. Brewing potions ||| Create Potion Post request
Handle POST requests at the /potions endpoint. The request consists of a Potion just brewed, containing the Student ID, and the list of ingredients. If a Recipe with these Ingredients (in any order) already exists, the Potion is a replica. If no such Recipe exists, the potion is a discovery, so the Recipe must also be persisted:
a) There is an endpoint at /potions, where a Student can brew the Potion by sending a POST request.
b) The list of Ingredients is checked if it matches any Potion.
c)If the brew is a discovery, the Recipe is persisted with the list of Ingredients, the Student, and with a name generated from the Student's name.
d)The response contains the persisted Potion.