6. Exploratory brewing ||| Create Potion Get request
In the beginning, a POST request is sent to /potions/brew, containing the student's ID. At this point, the soon-to-be Potion must be persisted with the Student and the Status. The response must contain the properties of the Potion so far, including its id. After this, brewing is done by sending a PUT request at /potions/{potion-id}/add, containing one ingredient at a time. Every added Ingredient must be checked whether it is previously known. If it is not, it needs to be persisted. The response must contain the Potion object, including the Ingredients added so far. After adding an Ingredient, help can be received by sending a GET request to /potions/{potion-id}/help. This endpoint must return all Recipes that include the current Ingredients. Once the Potion contains five Ingredients, you must check whether it is a replica or a discovery.

a) By sending a POST request to /potions/brew a new Potion is generated containing the Student and the status of brewing (this requirement has been already requested earlier - commit f32d6eca - I assume that this is a requirement/task error).
b) By sending a PUT request to /potions/{potion-id}/add the Potion with the potion-id gets updated with the new Ingredient. The response contains the updated Potion object.
c) If the potion has less than five Ingredients, sending a GET request to /potions/{potion-id}/help, returns the Recipes that contain the same Ingredients as the Potion brewing.